Fix tests for post body
The original author appears to have broken these tests in 958ee3b, when
`encode_options` was changed to only output a single value instead of
all values.

I'm not sure why this change was made nor why the tests were not
updated. It seems like you'd want a JSON formatted body, but on the
other hand, I couldn't find an ESI endpoint that used the body for
anything.

I'm fixing the tests to match the output for now and we'll see if I can
figure out what the intended behavior is here as I go on.